requirements.txt
asgiref==3.6.0
crispy-bootstrap4==2022.1
Django==4.2
django-crispy-forms==2.0
django-tinymce==3.6.1
Pillow==9.5.0
sqlparse==0.4.4
tzdata==2023.3
